Chore: Splits tests from implementation
