+++ kingdom/access/authorization/dsl.py
@@ -21,69 +21,12 @@ dsl.py
 
  There are some conditions to selectors:
    1. All selector must **always** be alone.
+
+OBS.: This is in WIP and should be thoroughly simplified & documented.
 """
 
 import string
 from typing import List
-
-valid_conditionals = [
-    "resource.id == 128f12334hjg || resource.id == 12839712893791823",
-    "resource.id==128f12334hjg||resource.id==12839712893791823",
-    "resource.id==*",
-    "resource.id ==*",
-    "resource.id== *",
-]
-
-simplified_conditionals = {
-    # This meanings that list[0] should be translated to list[1]
-    "selector conditional": [
-        "resource.id==*||resource.id==8123798fcd89||resource.id==8197498127",
-        "resource.id == *",
-    ]
-}
-
-invalid_conditionals = [
-    "resource.id == 9284234 ||",
-    "resource.id = 1238912fd || resource.id ==182789123",
-    "resource.id 12381723",
-    "resource.id == 1928309182 resource.id = 12930918",
-    "resource.id == 81927398123 || resource.id = 9128398",
-    "resource .id == 12893789123",
-    "resource. id == 89071239",
-    "resource.id == 182937 12893 827381723 || resource.id == 1892 9283"
-    "resource.id == 18297389f|resource.id==1827398f",
-    "resouce.id == 1283971283ff || subject.id == 1891273987123",
-    "resource.created_at == 2319833012707 || resource.id == faf76bc7",
-    "resource.id == 8129370192ff || resource.id == 8123091283908",
-    "resource.id!=291f767d8bc",
-]
-
-
-def test_conditionals_split():
-    input = [
-        "expr || newexpr",
-        "expr||newexpr",
-        "  expr || newexpr    ",
-        "expr|newexpr",
-        "expr   || newexpr||neewexpr||",
-        "ex pr   || new expr ||  ",
-        "  expr ||| newexpr    ",
-        "  expr|||newexpr    ",
-    ]
-
-    want = [
-        ("expr", "newexpr"),
-        ("expr", "newexpr"),
-        ("expr", "newexpr"),
-        False,
-        False,
-        False,
-        False,
-        False,
-    ]
-
-    got = [conditionals_split(sequence) for sequence in input]
-    assert got == want
 
 
 def conditionals_split(sequence: str):
@@ -106,29 +49,6 @@ def conditionals_split(sequence: str):
     return tuple(parsed_expr)
 
 
-def test_parse_identifier():
-    input = [
-        ".id",
-        "resource.id",
-        "   resourceeee.id",
-        "r es ource.id",
-        "resource .id",
-        " r esource.id",
-        "resource..id",
-    ]
-    want = [
-        (False,),
-        ("resource", ".id"),
-        ("resourceeee", ".id"),
-        (False,),
-        (False,),
-        (False,),
-        ("resource", "..id"),
-    ]
-    got = [parse_identifier(i) for i in input]
-    assert got == want
-
-
 def parse_identifier(expression):
     # since it's the beginning, we can strip this, to avoid overly
     # complicated iterations :-)
@@ -141,43 +61,6 @@ def parse_identifier(expression):
             if token == "." and len(identifier) > 0:
                 return identifier, expression[idx:]
             return (False,)
-
-
-def test_parse_reference():
-    input = [
-        ".",
-        ".id==",
-        " .id>",
-        " id==",
-        ".id   === ",
-        "..id..",
-        ".nope    =",
-        ".that!=",
-        "[index]>=",
-        ".fine0  ==",
-        ".name@ <=",
-        ".n a m e ===",
-        ".name >>",
-    ]
-
-    want = [
-        (False,),
-        ("id", "=="),
-        (False,),
-        (False,),
-        ("id", "=== "),
-        (False,),
-        ("nope", "="),
-        ("that", "!="),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-        ("name", ">>"),
-    ]
-
-    got = [parse_reference(ref) for ref in input]
-    assert got == want
 
 
 VALID_OPS = {"==", ">", "<", ">=", "<=", "!="}
@@ -222,35 +105,6 @@ def parse_reference(reference_expr):
             return (False,)
 
 
-def test_parse_identifier_reference():
-    input = [
-        "resource.id==",
-        "resourceeee.id>=",
-        "r esource.id  ==",
-        "resource .id <=",
-        " r esource.id =",
-        "resource.id  <<",
-        "resource.name <=",
-        "resource.na me !=",
-        "resource.name ===",
-        "rsrc!name ==",
-    ]
-    want = [
-        ("RESOURCE", "ID"),
-        ("RESOURCEEEE", "ID"),
-        False,
-        False,
-        False,
-        ("RESOURCE", "ID"),
-        ("RESOURCE", "NAME"),
-        False,
-        ("RESOURCE", "NAME"),
-        False,
-    ]
-    got = [parse_identifier_reference(expr) for expr in input]
-    assert got == want
-
-
 def parse_identifier_reference(expression):
     identifier, *reference_expr = parse_identifier(expression)
     if identifier is False:
@@ -262,66 +116,21 @@ def parse_identifier_reference(expression):
     return (identifier.upper(), reference.upper())
 
 
-def test_parse_operator():
-    input = [
-        "== 'd8f7s9d8f7'",  # valid
-        "== '8dfs8d7f9'",  # valid
-        "== ''",  # valid
-        "== \"'21f90912'\"  ",  # invalid
-        "== '21f90912'  ",  # valid
-        "= '*' ",  # invalid
-        '< "ddx"',  # invalid
-        "  < '3030fk30'",  # valid
-        " >    ''",  # valid
-        " >>    ''",  # invalid
-        "=== '*'",  # invalid
-        ">= '2fd04'",  # valid
-        "!= '*'",  # valid
-        " + '*'",  # invalid
-        "-- 'd'",  # invalid
-        "/   'xxvc'",  # invalid
-        "=+ '*'",  # invalid
-        "=- '*'",  # invalid
-    ]
-    want = [
-        ("==", "'d8f7s9d8f7'"),  # valid
-        ("==", "'8dfs8d7f9'"),  # valid
-        ("==", "''"),  # valid
-        (False,),  # invalid
-        ("==", "'21f90912'"),  # valid
-        (False,),  # invalid
-        (False,),  # invalid
-        ("<", "'3030fk30'"),  # valid
-        (">", "''"),  # valid
-        (False,),  # valid
-        (False,),  # invalid
-        (">=", "'2fd04'"),  # valid
-        ("!=", "'*'"),  # valid
-        (False,),  # invalid
-        (False,),  # invalid
-        (False,),  # invalid
-        (False,),  # invalid
-        (False,),  # invalid
-    ]
-    got = [parse_operator(op) for op in input]
-    assert got == want
-
-
 def parse_operator(operator_expr):
     operator_expr = operator_expr.strip()  # just to make sure
-    VALID_OPS = {"==", ">", "<", ">=", "<=", "!="}
-    VALID_TOKEN = {token for operator in VALID_OPS for token in operator}
 
     operator = ""
     parsing_idx = -1
+    # First we parse a known operator.
     for idx, token in enumerate(operator_expr):
-        if token in VALID_TOKEN:
+        if token in VALID_OPS_TOKEN:
             # the only valid thing that is being parsed
             operator += token
         else:
             parsing_idx = idx
             break
 
+    # Then we check if the rest is OK.
     rest = operator_expr[parsing_idx:]
     for idx, token in enumerate(rest):
         if token == "'":
@@ -338,52 +147,6 @@ def parse_operator(operator_expr):
             return (False,)
 
 
-def test_parse_valid_selector():
-    input = [
-        "   '128f12334hjg'",
-        "   '128f12334hjg'",
-        "'12839712893791823'",
-        " '12839712893791823'",
-        " '128f12334hjg'",
-        "'*'",
-        " '*'",
-        " '**'",
-        " '!'",
-        "'12837891ff",
-        "'123123'123123'",
-        "''",
-        "'fff\"dddsd'",
-        "'''",
-        "';1234234fgds00x;;",
-        "'f5f65b65c!!'",
-        "f4'dfadf7'",
-    ]
-
-    want = [
-        ("128f12334hjg",),
-        ("128f12334hjg",),
-        ("12839712893791823",),
-        ("12839712893791823",),
-        ("128f12334hjg",),
-        ("*",),
-        ("*",),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-        (False,),
-    ]
-
-    got = [parse_selector(cond) for cond in input]
-
-    assert got == want
-
-
 def parse_selector(selector_expr):
     ALL_TOKEN = "*"
     selector_expr = selector_expr.strip()
@@ -393,6 +156,7 @@ def parse_selector(selector_expr):
     def isselector(token):
         return token.isnumeric() or token.isidentifier() or token == ALL_TOKEN
 
+    # First we try to find a selector.
     for idx, token in enumerate(selector_expr):
         if idx == 0:
             # first token
@@ -409,7 +173,7 @@ def parse_selector(selector_expr):
         else:
             return (False,)
 
-    # are we dealig with an *?
+    # Edge case: are we dealig with an *?
     if ALL_TOKEN in selector and selector != "*":
         # plain comparison
         return (False,)
@@ -419,53 +183,6 @@ def parse_selector(selector_expr):
         # we shouldn't have anything left
         return (False,)
     return (selector,)
-
-
-def test_parse_expressions():
-    valid_input = [
-        "resource.id=='d8s7f987sdf'",
-        "resource.id == 'd8s7f987sdf'",
-        "resource.id == '*'",
-        "product.id=='*'",
-        "subject.salary > '1800'",
-        "subject.salary <= '1800'",
-        "some.name == 'ab9f8d0'",
-    ]
-    got = [parse_expression(expr) for expr in valid_input]
-    want = [
-        ("resource", "id", "==", "d8s7f987sdf"),
-        ("resource", "id", "==", "d8s7f987sdf"),
-        ("resource", "id", "==", "*"),
-        ("product", "id", "==", "*"),
-        ("subject", "salary", ">", "1800"),
-        ("subject", "salary", "<=", "1800"),
-        ("some", "name", "==", "ab9f8d0"),
-    ]
-    assert got == want
-
-    invalid_input = [
-        "resource..id == '8d9f7a8f'",
-        ".id == 'xxx'",
-        "'resource'.id == '*'",
-        "reso urce.id == '*'",
-        "resource.id === '*'",
-        "resource.id='dgv8bf'",
-        'resource.id == "*"',
-        "subject.salary > 1800",
-    ]
-
-    got = [parse_expression(expr) for expr in invalid_input]
-    want = [
-        False,
-        False,
-        False,
-        False,
-        False,
-        False,
-        False,
-        False,
-    ]
-    assert got == want
 
 
 def parse_expression(expr):
